+++ Samples/Bank/Web/API/accounts/debit/DebitAccount.ts
@@ -2,10 +2,11 @@
  *  **DO NOT EDIT** - This file is an automatically generated file.
  *--------------------------------------------------------------------------------------------*/
 
+import { Nullable } from './Nullable';
 
 export type DebitAccount = {
     id: string;
     name: string;
     owner: string;
-    balance: number;
+    balance: Nullable;
 };