+++ Source/Extensions/Orleans/Configuration/EndPoint.cs
@@ -1,0 +1,22 @@
+// Copyright (c) Aksio Insurtech. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Net;
+
+namespace Aksio.Cratis.Extensions.Orleans.Configuration;
+
+/// <summary>
+/// Represents the configuration of an <see cref="IPEndPoint"/>.
+/// </summary>
+public class EndPoint
+{
+    /// <summary>
+    /// Gets the IP address.
+    /// </summary>
+    public string Address { get; init; } = string.Empty;
+
+    /// <summary>
+    /// Gets the port.
+    /// </summary>
+    public int Port { get; init; } = 30000;
+}