+++ Source/Kernel/Projections/Engine/Projection.cs
@@ -113,7 +113,7 @@ public class Projection : IProjection
     public void SetEventTypesWithKeyResolvers(IEnumerable<EventTypeWithKeyResolver> eventTypesWithKeyResolver)
     {
         EventTypesWithKeyResolver = eventTypesWithKeyResolver;
-        EventTypes = eventTypesWithKeyResolver.Select(_ => _.EventType);
+        EventTypes = eventTypesWithKeyResolver.Select(_ => _.EventType).ToArray();
         _eventTypesToKeyResolver = eventTypesWithKeyResolver.ToDictionary(_ => _.EventType, _ => _.KeyResolver);
     }