+++ Source/Kernel/Server/Program.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.Net;
 using Aksio.Cratis.Compliance;
 using Aksio.Cratis.Compliance.MongoDB;
 using Aksio.Cratis.Events.Projections;
@@ -14,7 +13,6 @@ using Aksio.Cratis.Events.Store;
 using Aksio.Cratis.Events.Store.MongoDB;
 using Aksio.Cratis.Execution;
 using Orleans;
-using Orleans.Configuration;
 using Orleans.Hosting;
 using Serilog;
 
@@ -30,34 +28,27 @@ public static class Program
         return CreateHostBuilder(args).RunConsoleAsync();
     }
 
-    public static IHostBuilder CreateHostBuilder(string[] args) =>
-         Host.CreateDefaultBuilder(args)
-            .UseAksio(MicroserviceId.Unspecified, _ => _.InSilo())
-            .UseOrleans(_ => _
-                .UseLocalhostClustering()
-                .ConfigureServices(_ => _
-                    .AddSingleton<IProjectionPositions, MongoDBProjectionPositions>()
-                    .AddSingleton<IChangesetStorage, MongoDBChangesetStorage>()
-                    .AddSingleton<IEncryptionKeyStore>(sp => new CacheEncryptionKeyStore(sp.GetService<MongoDBEncryptionKeyStore>()!))
-                    .AddSingleton<ISchemaStore, MongoDBSchemaStore>()
-                    .AddSingleton<IEventLogStorageProvider, MongoDBEventLogStorageProvider>()
-                    .AddSingleton<IProjectionDefinitionsStorage, MongoDBProjectionDefinitionsStorage>()
-                    .AddSingleton<IProjectionPipelineDefinitionsStorage, MongoDBProjectionPipelineDefinitionsStorage>()
-                    .AddSingleton<IProjectionDefinitionsStorage, MongoDBProjectionDefinitionsStorage>())
-                .Configure<EndpointOptions>(options =>
-                {
-                    options.SiloPort = 11111;
-                    options.SiloListeningEndpoint = new IPEndPoint(IPAddress.Any, 11111);
-                    options.GatewayPort = 30000;
-                    options.GatewayListeningEndpoint = new IPEndPoint(IPAddress.Any, 30000);
-                })
-                .AddConnectedClientsTracking()
-                .AddEventSequenceStream()
-                .UseMongoDBReminderService()
-                .AddSimpleMessageStreamProvider(WellKnownProviders.ObserverHandlersStreamProvider, cs => cs.Configure(o => o.FireAndForgetDelivery = false))
-                .AddExecutionContext())
-            .ConfigureWebHostDefaults(_ => _
-                .UseStartup<Startup>());
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+             Host.CreateDefaultBuilder(args)
+                .UseAksio(_ => _.InSilo())
+                .UseOrleans(_ => _
+                    .UseCluster()
+                    .ConfigureServices(_ => _
+                        .AddSingleton<IProjectionPositions, MongoDBProjectionPositions>()
+                        .AddSingleton<IChangesetStorage, MongoDBChangesetStorage>()
+                        .AddSingleton<IEncryptionKeyStore>(sp => new CacheEncryptionKeyStore(sp.GetService<MongoDBEncryptionKeyStore>()!))
+                        .AddSingleton<ISchemaStore, MongoDBSchemaStore>()
+                        .AddSingleton<IEventLogStorageProvider, MongoDBEventLogStorageProvider>()
+                        .AddSingleton<IProjectionDefinitionsStorage, MongoDBProjectionDefinitionsStorage>()
+                        .AddSingleton<IProjectionPipelineDefinitionsStorage, MongoDBProjectionPipelineDefinitionsStorage>()
+                        .AddSingleton<IProjectionDefinitionsStorage, MongoDBProjectionDefinitionsStorage>())
+                    .AddConnectedClientsTracking()
+                    .AddEventLogStream()
+                    .UseMongoDBReminderService()
+                    .AddSimpleMessageStreamProvider("observer-handlers", cs => cs.Configure(o => o.FireAndForgetDelivery = false))
+                    .AddExecutionContext())
+                .ConfigureWebHostDefaults(_ => _
+                    .UseStartup<Startup>());
 
     static void UnhandledExceptions(object sender, UnhandledExceptionEventArgs args)
     {